refactor(tests): split matcher helpers into matchDefinitionForAssignment.test-utilities.ts, add expectNoMatchWithFixture, collapse 5 no-match cases (-130 LOC)
